requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.0.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
requests==2.32.3
pydantic==2.9.2
numpy==1.26.4
ffmpeg-python==0.2.0
openai-whisper==20240930